requirements.txt
asgiref==3.9.1
brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cssselect2==0.8.0
django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
fonttools==4.60.1
idna==3.10
pillow==12.0.0
psycopg2-binary==2.9.10
pycparser==2.23
pydyf==0.11.0
pyjwt==2.10.1
pyphen==0.17.2
python-dotenv==1.1.1
qrcode==8.2
reportlab==4.4.4
requests==2.32.5
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
